Rename to PageBuilderUtil, add test
