+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/valueset/ConceptManager.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 import gov.cdc.nbs.questionbank.entity.CodeValueGeneral;
 import gov.cdc.nbs.questionbank.entity.CodeValueGeneralRepository;
-import static gov.cdc.nbs.questionbank.util.Util.requireNotEmpty;
-import static gov.cdc.nbs.questionbank.util.Util.requireNonNull;
+import static gov.cdc.nbs.questionbank.util.PageBuilderUtil.requireNotEmpty;
+import static gov.cdc.nbs.questionbank.util.PageBuilderUtil.requireNonNull;
 import gov.cdc.nbs.questionbank.valueset.command.ConceptCommand;
 import gov.cdc.nbs.questionbank.valueset.exception.ConceptNotFoundException;
 import gov.cdc.nbs.questionbank.valueset.exception.DuplicateConceptException;
